Refactored log in form into its own component
